requirements.txt
python-telegram-bot==20.7  
polars==0.20.2
python-dotenv==1.0.0